changes to configurations

+++ app/Http/Controllers/Property/PropertyRequestController.php
@@ -30,29 +30,45 @@ class PropertyRequestController
 
     public function store(CreatePropertyRequest $request)
     {
-        PropertyRequest::create($request->validated());
+        $user = auth()->user();
+        $user->propertyRequests()->create($request->validated());
+
+        return redirect()->route('property.requests.index')->with('success', 'Kerkesa e prones u shtua!');
     }
 
-    public function edit($id)
+    public function edit(PropertyRequest $propertyRequest)
     {
-        $propertyRequest = PropertyRequest::findOrFail($id);
+        if(auth()->id() !== $propertyRequest->user_id) {
+            abort(403);
+        }
         return Inertia::render('user/PropertyRequestEdit', [
             'propertyRequest' => $propertyRequest,
         ]);
     }
 
-    public function update(CreatePropertyRequest $request, $id)
+    public function update(CreatePropertyRequest $request, PropertyRequest $propertyRequest)
     {
-        $propertyRequest = PropertyRequest::findOrFail($id);
+        if(auth()->id() !== $propertyRequest->user_id) {
+            abort(403);
+        }
         $propertyRequest->update($request->validated());
 
-        return redirect()->route('properties.index')->with('success', 'Property request updated successfully.');
+        return redirect()->route('property.requests.index')->with('success', 'Property request updated successfully.');
     }
 
-    public function delete($id)
+    public function destroy(PropertyRequest $propertyRequest)
     {
-        $propertyRequest = PropertyRequest::findOrFail($id);
+        if(auth()->id() !== $propertyRequest->user_id) {
+            abort(403);
+        }
         $propertyRequest->delete();
-        return redirect()->route('properties.index')->with('success', 'Property request deleted successfully.');
+        return redirect()->route('property.requests.index')->with('success', 'Property request deleted successfully.');
+    }
+
+    public function show(PropertyRequest $propertyRequest)
+    {
+        return Inertia::render('user/ViewPropertyRequest', [
+            'propertyRequest' => $propertyRequest
+        ]);
     }
 }
